backend/requirements.txt
tensorflow==2.15.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
statsmodels==0.14.1
pandas==2.2.0
scikit-learn==1.4.0
transformers==4.37.2
tf-keras

